[Xamarin.Android.Build.Tasks] ILRepack if `NuGet.ProjectModel` Exists (#2642)
We saw the following error in the latest versions of monodroid:

	/Library/Frameworks/Mono.framework/External/xbuild/Xamarin/Android/Xamarin.Android.Common.targets(1928,2): error : Could not load file or assembly 'NuGet.ProjectModel, Version=4.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35' or one of its dependencies.

After investigation it turns out that the
`Xamarin.Android.Build.Tasks.csproj` project was being run multiple
times (as expected).  However the `CopyFilesToOutputDirectory` target
was being run on the second build, which ran an internal
`_CopyFilesMarkedCopyLocal` target on the second run.

The result of both of these targets was to copy over the old assembly
over the `ILRepack`ed version and copy over ALL the NuGet assemblies
from the NuGet packages cache.

This then turns into the above error at build time because the system
has a different version of NuGet.  This was why we were using
`ILReapck` in the first place.

We tried a number of different solutions:

We tried to update the file within `$(IntermediateOutputPath)` which
looked like it would solve the problem initially.  However in that
instance the NuGet files were still copied over to `$(OutputPath)`
and might cause problems in the final release.  There was also
another problem with `ILRepack`: if we fixed up the file in the
`$(IntermediateOutputPath)` before the `CopyFilesToOutputDirectory`
target ran, we could end up with `ILRepack` running again on the same
assembly.  This causes it to pull in ALL the NuGet stuff again and we
end up with a 5MB file rather than a 3MB one... Which was weird.

The simplest solution is to run `ILRepack` after the
`CopyFilesToOutputDirectory` target *only if* the
`NuGet.ProjectModel` assembly exists in `$(OutputPath)`.  Since the
`ILRepacker` target is the one that deletes the NuGet assemblies we
can be sure that if they exist we need to fix up the main assembly.
While this might mean that ILRepack runs twice in a build, it does
mean we get consistent output.