[Xamarin.Android.Build.Tasks] Fix ILRepack to use `AfterBuild`.
When we changed over to use `CopyFilesToOutputDirectory` for the
`AfterTargets` value we forgot that we are still using `xbuild`
for building this branch. This results in the following warning

	Xamarin.Android.Build.Tasks.csproj:  warning : Target 'CopyFilesToOutputDirectory', not found in the project

which means the ILRepack does not run.

So this PR moves back to using `AfterBuild` for that value. This
should allow the `Xamarin.Android.Build.Tasks` assembly to be
properly repacked.