[Xamarin.Android.Build.Tasks] Fix Ndk .exe lookup.
The r19 Ndk on windows does not use `.exe` files
for some of its compiler tools. It uses `.cmd`.
As a result our current setup will not find these
files and you end up with errors like

	error XA5101: Toolchain utility 'armv7a-linux-androideabi16-clang.exe' for target Arm was not found. Tried in path: "C:\Users\dlab14\android-toolchain\ndk\toolchains\llvm\prebuilt\windows-x86_64"

This commit changes the code to use the
`MonoAndroidHelper.GetExecutablePath` method which
uses the `PATHEXT` environment variable to search
for various exe file extensions.