[NDK,Windows] Fix NDK r19 clang scripts bug on Windows (#2664)
The llvm toolchain directory contains a selection of shell scripts
(both Unix `#!/sh` and Windows `.cmd`) which call `clang/clang++`
with different `-target` parameters depending on both the target
architecture and API level.  For example, the clang/clang++ compilers
targeting `aarch64` on API-28 will have the following Unix shell
scripts present in the toolchain `bin` directory:

	aarch64-linux-android28-clang
	aarch64-linux-android28-clang++

However, the Windows version of the NDK has a bug where there is onl
one  Windows counterpart to the above Unix scripts:

	aarch64-linux-android28-clang.cmd

This script, despite its name, suggesting that it calls `clang.exe`
when in fact it calls `clang++.exe`, which breaks compilation of some
C programs (including the code generated by Mono's `mkbundle.exe`
utility) because `clang++` treats the input as C++.  There is no
corresponding `aarch64-linux-android28-clang++.cmd`, so invocation of
`clang.exe` becomes harder and, most certainly, non-standard as far
as cross-platform NDK compatibility is concerned.

This commit tries to rectify the situation by special-casing the
compiler tool handling to return path to the actual `.exe` instead of
the `.cmd` and provide the necessary target parameters on the
compiler invocation.