[timing] don't use `git clone --depth=1` (#2672)
Context: https://github.com/Microsoft/SmartHotel360-Mobile

Downstream in monodroid we building the SmartHotel360 app as part of
our MSBuild project build timing collection.

The problem is with following set of git commands:

	git clone https://github.com/Microsoft/SmartHotel360-Mobile.git --depth=1
	git checkout ad41901761df12802ab332d78b9eedb3248f3f1e

If Microsoft/SmartHotel360-Mobile@ad419017 is not the latest commit,
this command will fail:

	fatal: reference is not a tree: ad41901761df12802ab332d78b9eedb3248f3f1e

I inadvertently broke this by updating our fork to the latest
Microsoft/SmartHotel360-Mobile/master commit.  Whoops!

Some options to fix this would be:

 1. Create a branch, then don't every touch it! Use `--branch` during
    the `git clone`.
 2. Don't use `--depth=1` at all.

Looking into it `--depth=1` is not saving us that much on this repo:
it has around 300 commits now and using `--depth=1` only saves 16MB
of disk space.

Skipping the `--depth=1` is a lot simpler, and we don't have to worry
about future commits breaking anything.