[Xamarin.Android.Build.Tasks] NdkUtil tests; PATHEXT precedence (#2659)
Commit 06e920c1 introduced NDK r19 usage, which promptly broke the
Windows build:

	error XA5101: Toolchain utility 'armv7a-linux-androideabi16-clang.exe' for target Arm was not found. Tried in path: "C:\Users\dlab14\android-toolchain\ndk\toolchains\llvm\prebuilt\windows-x86_64"

This was subsequently fixed in 08c986e7.

This doesn't mean that we can't improve matters. :-)

Add an `NdkUtilTests` test suite to test the behavior of the
`NdkUtil` class, and ensure that we can find the compiler toolchain
utilities that are required.

Update the semantics of `MonoAndroidHelper.Executables()` so that the
"source" executable is returned *last*, instead of *first*.

The intention behind `MonoAndroidHelper.Executables()` was to allow
searching via program *base names*: if a utility is named `example`
on Linux/macOS but is named `example.exe` on Windows,
`MonoAndroidHelper.Executables("example")` should return
`example.exe` on Windows.

...which is the case, *so long as* there is no `example` file on
Windows either.

Enter the NDK r19 on Windows, which has:

	toolchains\llvm\prebuilt\windows-x86_64\bin\x86_64-linux-android26-clang
	toolchains\llvm\prebuilt\windows-x86_64\bin\x86_64-linux-android26-clang.cmd

Meaning `MonoAndroidHelper.GetExecutablePath(@"…\bin", "x86_64-linux-android26-clang")`
returns a *Unix shell script* instead of the Windows `.cmd`!

Now, this is in fact *no longer relevant* because commit 08c986e7
fixes things so that we don't even use the `.cmd` file -- because the
`*clang.cmd` file for certain API levels invokes `clang++.exe`, not 
`clang.exe`, which is another level of whack -- but the underlying
semantic question remains: Given the NDK r19 directory structure,
which file should `MonoAndroidHelper.Executables()` return first?

It *should* return `x86_64-linux-android26-clang.cmd`.  It actually
returns `x86_64-linux-android26-clang`.

Fix `MonoAndroidHelper.Executables()` so that `%PATHEXT%` is checked
*first*, *before* returning the original (unmodified) filename.
This will allow `Executables("x86_64-linux-android26-clang")` to
generate `x86_64-linux-android26-clang.cmd` first, which will give
the `.cmd` file *precedence*, which is desirable *in general*.