[docs] Reorganize build instructions by OS
Context: https://gitter.im/xamarin/xamarin-android?at=5c4a25ba7a0f4d5b1900f7a4

Move the build dependencies and build instructions for macOS and Linux
into one subdirectory, and move the build instructions for Windows into
another.  This is a preparatory step for a follow-up commit that will
add a Windows-specific build dependencies page.

The motivation for these changes comes from the linked Gitter comment,
where a user found that trying to use Mono to build
`Xamarin.Android.sln` on Windows caused build errors.  This result
highlighted a problem that the Windows build instructions currently link
to the macOS and Linux build dependencies page, and that page lists Mono
as a build dependency, when Mono is *not* in fact a build dependency for
Windows.  A possible fix could be to add a Windows section to the build
dependencies page, but splitting the Windows build dependencies into
their own page will hopefully do an even better job of making the build
steps for Windows easy to follow.