Drop MonoPackageManager for APIs older than 9
Context: https://github.com/xamarin/xamarin-android/issues/2690
Fixes: https://github.com/xamarin/xamarin-android/issues/2690

Xamarin.Android and Mono are no longer capable of running on API levels lower
than 9 (if built with NDK r14) or 16 (if built with NDK r18+). This commit drops
the version of `MonoPackageManager` targetting API levels < 9.

If an application with its minimum target sdk set to anything < 9 is build it
results in a compilation error when building Java code:

    error :  error: method init in class Runtime cannot be applied to given types; [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
    error :	Runtime.init ( [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
    error :   required: String,String[],String,String[],ClassLoader,String[],String[],String,int,String[] [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
    error :   found: String,String[],String,String[],ClassLoader,String[],String[],String [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
    error :   reason: actual and formal argument lists differ in length [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]

This is caused by the changes introduced in the runtime performance commit
https://github.com/xamarin/xamarin-android/commit/b90d3ab717cbfd38b2e6a307b4387d2d6cf04af8
which changed the `Runtime.init` signature but *only* in the version of
`MonoPackageManager` targetting the currently supported minimum SDK levels.