[Xamarin.Android.Build.Tasks] Drop MonoPackageManager.api4.java (#2691)
Xamarin.Android and Mono are no longer capable of running on API
levels lower than 9 (if built with NDK r14) or 16 (if built with
NDK r18+).  This commit drops the version of `MonoPackageManager`
targeting API levels < 9.

If an application with its minimum target sdk set to anything < 9 is
built it results in a compilation error when building Java code:

	error :  error: method init in class Runtime cannot be applied to given types; [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
	error :	Runtime.init ( [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
	error :   required: String,String[],String,String[],ClassLoader,String[],String[],String,int,String[] [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
	error :   found: String,String[],String,String[],ClassLoader,String[],String[],String [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]
	error :   reason: actual and formal argument lists differ in length [C:\xms\NPlus1DaysOfMvvmCross\N-04-ValueConverters\Value.Droid\Value.Droid.csproj]

This is caused by the changes introduced in the runtime performance
commit b90d3ab7 which changed the `Runtime.init()` signature but
*only* in the version of `MonoPackageManager` targeting the currently
supported minimum SDK levels.