[Xamarin.Android.Build.Tasks] Provide file.encoding for dx tool
Fixes #2685

In commit 8571a7f we switched the `CompileToDalvik` task over
to use the `--input-list` argument. This helped reduce the
overall length of the command line arguments passed. It
had a side effect of causing a problem if the path
contained non-ASCII characters. We end up with the
following error

	java.io.FileNotFoundException: C:\Temp\Scratch.├£ml├ñ├╝ts\Scratch.├£ml├ñ├╝ts\obj\Debug\90\android\bin\classes.zip (The system cannot find the path specified)

This is because while the `input-list` file is fine,
the paths within it might contain non-ASCII characters.
So we need to let java know that, we can do that via the

	file.encoding=UTF8

argument. `CompileToDalvik` can be run in two modes, the first
is just invoking the `dx` executable. The second is invoking
the `dx.jar` via `java`. Both of these need to be supported.

For the normal `java` invocation we can just pass the argument
as normal via `-Dfile.encoding=UTF8`. However for `dx` that is
not a valid argument. We instead need to provide it via

	-JDfile.encoding=UTF8

The `-J` tells `dx` that this is a `java` argument. `dx` will then
take everything after the `-J` , prepend a `-` and pass that to
`java`.

One of the unit tests has been updated to include non-ASCII
characters.