[Xamarin.Android.Build.Tasks] Provide file.encoding for dx tool (#2695)
Fixes: http://work.devdiv.io/784551
Fixes: https://developercommunity.visualstudio.com/content/problem/437807/v-isual-studio-2019-preview-20.html
Fixes: https://github.com/xamarin/xamarin-android/issues/2685

Context: https://github.com/xamarin/xamarin-android/pull/2688#issuecomment-460887811
Context: https://github.com/xamarin/xamarin-android/pull/2688#issuecomment-460893374

In commit 8571a7f we updated the `<CompileToDalvik/>` task to
generate and use a "response file" via the `--input-list` argument.
This helped reduce the overall length of the command line arguments
passed.  It had a side effect of causing a problem if the path
contains non-ASCII characters, causing the following error on Windows:

	java.io.FileNotFoundException: C:\Temp\Scratch.├£ml├ñ├╝ts\Scratch.├£ml├ñ├╝ts\obj\Debug\90\android\bin\classes.zip (The system cannot find the path specified)

This is because while the `input-list` file itself is fine, the paths
within it might contain non-ASCII characters.  We need to let java
know that, which we can do that via the `java` argument:

	-Dfile.encoding=UTF8

The `<CompileToDalvik/>` task can be run in two modes, the first is
just invoking the `dx` executable, and the second is invoking
`dx.jar` via `java`.  Both of these need to be supported.

For the normal `java` invocation we can just pass the argument with
`-Dfile.encoding=UTF8`.  However for `dx` that is not a valid
argument.  We instead need to provide it via:

	-JDfile.encoding=UTF8

The `-J` tells `dx` that this is a `java` argument.  `dx` will then
take everything after the `-J` , prepend a `-` and pass that to
`java`.

Finally, there was a [problem of Unicode Normalization Forms][0]
and `string.Replace()`.  On macOS, it is possible within the
`<AndroidComputeResPaths/>` task for `$(ProjectDir)` to use a
different filesystem normalization than `@(AndroidAsset)`, because
`@(AndroidAsset)` comes from a directory traversal and will use
Unicode Normalization Form D [(-ish)][1] on HFS+, while
`$(ProjectDir)` will come from the `msbuild` command-line invocation
performed by the unit tests and will be in Unicode Normalization
Form C, and through "luck" `string.Replace()` doesn't handle this:

	var e_acute = "\u00e9";          // http://www.fileformat.info/info/unicode/char/e9/index.htm
	var e_formc = e_acute.Normalize (System.Text.NormalizationForm.FormC);
	var e_formd = e_acute.Normalize (System.Text.NormalizationForm.FormD);
	e_formc.StartsWith (e_formd);    // true
	e_formc.Replace (e_formd, "");   // == "é" == e_formc; no replacement performed!

`<AndroidComputeResPaths>` uses `string.Replace()` to convert a "full"
path name into a "relative" path name, but as the `string.Replace()`
didn't do anything here with the mismatched Unicode Normalization,
`<AndroidComputeResPaths/>` would return *lowercased full paths* to
subsequent tasks, not the expected lowercased *relative* paths, which
ultimately results in a `Resource.designer.cs` which is *empty*:

	Task "AndroidComputeResPaths" (TaskId:62)
	    ...
	    IntermediateFiles: (TaskId:62)
	      /users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/bin/testrelease/temp/buildreleaseappwitha initandümläüts(truedx)/resources/drawable-hdpi/icon.png (TaskId:62)
	      ... wat?! ...
	...
	MainActivity.cs(22,20,22,28): error CS0103: The name 'Resource' does not exist in the current context

Fix this problem by normalizing the strings into the same Unicode
Normalization Form *before* performing the `string.Replace()` call,
which works as expected/intended, restoring the semantics of
`<AndroidComputeResPaths>`.

Update one of the unit tests has been to use non-ASCII characters.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder-release/623/testReport/junit/Xamarin.Android.Build.Tests.BuildTest/BuildApplicationWithSpacesInPathAndÜmläüts/_True__dx________Release/
[1]: https://en.wikipedia.org/wiki/HFS_Plus#Design