[Xamarin.Android.Build.Tasks] use NDKMinimumApiAvailable for minSdkVersion warning
Context: https://github.com/xamarin/xamarin-android/pull/2698

In #2698, we added a new build warning for projects using a
`//uses-sdk/@android:minSdkVersion` lower than Xamarin.Android
supports. This change checking for API-9 was appropriate for 16.0, but
not 16.1/master.

We should instead use `XABuildConfig.NDKMinimumApiAvailable` to decide
to show these warnings or not. I updated the relevant test &
warning code documentation about this as well.

~~ Other changes ~~

While using `XABuildConfig` in `Xamarin.Android.Build.Tests`, I
noticed a compiler warning. It seems the type was found twice, since
it has a reference to `Xamarin.Android.Build.Tasks`. We can remove the
`<Compile/>` item for `XABuildConfig.cs` in this project.