[Xamarin.Android.Build.Tasks] Improve Android NDK validation errors
Fixes: https://github.com/xamarin/xamarin-android/issues/2708

Move the check for a `null` or empty `AndroidNdkDirectory` from
`MakeBundleNativeCodeExternal.Execute()` into
`NdkUtil.ValidateNdkPlatform()`, and update the error message.  Make
`ValidateNdkPlatform()` fail with that same error if it cannot get the
version of the NDK.  That way, the `MakeBundleNativeCodeExternal` and
`Aot` tasks will both hit that new updated error first before they try
to call any other methods from `NdkUtil`.  This solves the problems from
the linked issue.  In short, the two tasks were each producing a
different error message when invoked with an invalid or missing
`AndroidNdkDirectory`, and neither message suggested how to resolve the
error.