[Xamarin.Android.Build.Tasks] Correct Android NDK existence check
Fixes the test failures introduced in: https://github.com/xamarin/xamarin-android/pull/2709

I initially authored the changes in [#2709][0] against a `master` commit
that was a few before the [switch to Android NDK r19][1].  When I
locally merged the changes onto the tip of `master`, I didn't notice
that the `MakeBundleNativeCodeExternal` and `Aot` tasks were now calling
the new method `NdkUtil.Init()` before `NdkUtil.ValidateNdkPlatform()`.
And unfortunately since there was some urgency to get the corresponding
backport ready for the `d16-0` branch, I didn't get a chance to re-build
and re-test locally against the tip of `master`.  Long story short, to
support the new order of method calls, the NDK existence check now needs
to happen at the time of the call to `GetNdkToolchainRelease()` in
`NdkUtil.Init()`.

[0]: https://github.com/xamarin/xamarin-android/pull/2709
[1]: https://github.com/xamarin/xamarin-android/commit/06e920c18b71a165eded0a5782f1a285ef20a4bc