[tests] Make the Mono.Android_TestsMultiDex run
We see these errors in our jenkins builds:

    /Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/build-tools/scripts/TestApks.targets(217,5): error : Input file 'logcat-Release-Aot-Mono.Android_TestsMultiDex.txt' doesn't exist. [/Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/tests/RunApkTests.targets]
      Build continuing because "ContinueOnError" on the task "ProcessLogcatTiming" is set to "ErrorAndContinue".

Turns out, that the reason for that is, that the
`Mono.Android_TestsMultiDex` test is not run, because no
`RunInstrumentationTests` nor `RunUITests` is used as their condition
clauses are resolved as **false**.

This is fixed by adding `TestApkInstrumentation` (and
`TestApkPermission` plus `TestApkPermission`) item groups. This makes
the test run thru the `RunInstrumentationTests` task.