[Tests] Fix Layout so the test app actually runs
Currently when running the Mono.Android-Test app
we get the following error

	Java.Lang.RuntimeException: Binary XML file line #1: You must supply a layout_width attribute.

This is because the `ScrollView` now needs to have
the `layout` properties for `width` and `height`.
Filling those in we then get

	Scrollview can host only one direct child

So we then need to add a `LinearLayout` to the `ScrollView`
so that it removes that error as well.