[tests] Fix Layout so the test app actually runs (#2720)
Currently when running the `Mono.Android-Test` app we get the
following error:

	Java.Lang.RuntimeException: Binary XML file line #1: You must supply a layout_width attribute.

This is because `ScrollView` now needs to have the properties
`//@android:layout_width` and `//@android:layout_height`.  Once we
provide those, we then get:

	Scrollview can host only one direct child

We then need to add a `LinearLayout` to the `ScrollView` so that it
removes that error as well.