Fix issues found by static analysis + compiler warnings
This commit adds a number of measures to try to detect potential issues in the
native C++ code in Xamarin.Android.

Firstly, it passes a number of warning flags to the C++ compiler to enable
diagnostics disabled by default related to type conversion (signed vs unsigned
integers), sign and unsigned types comparison, use of potentially uninitialized
variables, use of deprecated APIs. These flags are enabled ONLY for our code in
`src/monodroid` since we can't do much about e.g. sqlite code and enabling them
for 3rd party code would merely result in useless noise during the build which
could hide issues in our code.

Secondly, it adds an MSBuild target named `RunStaticAnalysis` in `src/monodroid`
to run the `clang-check` utility which analyzes source code for each target we
build and saves the analysis report into a text file for investigation. This
target needs to be ran manually (perhaps we should add a step to our CI pipeline
at some point to run it) by the developers working on the native
code (recommended when non-trivial changes are made to the code)

The remainder of changes in this commit is the result of applying the above
measures - most of them fix issues related to integer sign.