[tests] more logging around MicrosoftNetSdkDirectory
Downstream in monodroid, we have a test failure:

    18:56:55   1) Failed : Xamarin.Android.Build.Tests.PackagingTest.NetStandardReferenceTest
    18:56:55   Microsoft.NET.Sdk not found.
    18:56:55   at Xamarin.Android.Build.Tests.PackagingTest.NetStandardReferenceTest ()

We likely introduced this in ef7a38d, but there isn't enough logging
to see what went wrong. When I logged into the build machine the
directory we are looking for looked correct.

I changed `PackagingTest` to log the directory it used, and added a
`Console.WriteLine` call in `Builder.cs`.